requirements.txt
Flask==0.12.2
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
click==6.7
itsdangerous==0.24
requests==2.10.0
gunicorn==19.6.0
googletrans
python-aiml